# This is a combination of 2 commits. # This is the 1st commit message:
Use vitest for testing the packages in this monorepo. Also setup the way
we can run this code on Remix example apps.
The way it works is you add the package in `/packages` folder. And if
you want a new remix app example add it to `/examples`. It's explained
in the README. Please follow the instructions

# This is the commit message #2:

Added workspace layout config to nx.json