@remix-storage/bodyparser In this PR we introduce the first version of bodyparser. Parsing a request with their fields and files and applying validations is crucial. With this code we want to provide to the final dev a helpful method of defining an expected output from a user/request, apply parsing on that request and validate it agains an Zod schema. Although also is posible to just use the parser without schema.
The problem with Remix `uploadHandler` for file uploads is that
it doesn't handle the removal of files when the request is not fully
valid. Ex.: I want a user to submit an email, a username and an avatar
file. If we use `uploadHandler` the file is uploaded. But if later in
the request we check that `email` doesn't have a valid format is the
developer the one responsible for `unlinking` the uploaded file because
is not going to be used